+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - audio.pl normalize has an option to re-evaluate site and node configurations (todo #169)
+    - audio.pl normalize checks the provided format (todo #168)
 
 ### 4.20.2